let's use functions
