+++ packages/services/schema/src/lib/trim-descriptions.ts
@@ -1,0 +1,31 @@
+import { visit, type ASTNode, type DocumentNode } from 'graphql';
+
+export function trimDescriptions(doc: DocumentNode): DocumentNode {
+  function trim<T extends ASTNode>(node: T): T {
+    if (node && 'description' in node && node.description) {
+      (node.description as any).value = node.description.value.trim();
+    }
+
+    return node;
+  }
+
+  return visit(doc, {
+    SchemaDefinition: trim,
+    ObjectTypeDefinition: trim,
+    ObjectTypeExtension: trim,
+    InterfaceTypeExtension: trim,
+    UnionTypeExtension: trim,
+    InputObjectTypeExtension: trim,
+    EnumTypeExtension: trim,
+    SchemaExtension: trim,
+    ScalarTypeExtension: trim,
+    FieldDefinition: trim,
+    InputValueDefinition: trim,
+    InterfaceTypeDefinition: trim,
+    UnionTypeDefinition: trim,
+    EnumTypeDefinition: trim,
+    EnumValueDefinition: trim,
+    InputObjectTypeDefinition: trim,
+    DirectiveDefinition: trim,
+  });
+}
